revert pc select to original
